Fix/user delete softdelete

+++ src/main/java/com/formssafe/domain/user/service/UserService.java
@@ -1,15 +1,18 @@
 package com.formssafe.domain.user.service;
 
+import com.formssafe.domain.form.service.FormService;
 import com.formssafe.domain.oauth.client.OauthMemberClientComposite;
 import com.formssafe.domain.user.dto.UserRequest.JoinDto;
 import com.formssafe.domain.user.dto.UserRequest.LoginUserDto;
 import com.formssafe.domain.user.dto.UserRequest.NicknameUpdateDto;
 import com.formssafe.domain.user.dto.UserResponse.UserProfileDto;
+import com.formssafe.domain.user.entity.OauthId;
 import com.formssafe.domain.user.entity.User;
 import com.formssafe.domain.user.repository.UserRepository;
 import com.formssafe.global.exception.type.BadRequestException;
 import com.formssafe.global.exception.type.DataNotFoundException;
 import com.formssafe.global.exception.type.ForbiddenException;
+import com.formssafe.global.util.CommonUtil;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,6 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
     private final UserRepository userRepository;
     private final OauthMemberClientComposite oauthMemberClientComposite;
+    private final FormService formService;
 
     @Transactional
     public void join(JoinDto request, LoginUserDto loginUser) {
@@ -39,9 +43,12 @@ public class UserService {
     }
 
     public UserProfileDto getProfile(LoginUserDto loginUser) {
-        Long userId = loginUser.id();
-        User user = userRepository.findById(userId).orElseThrow(() ->
-                new DataNotFoundException("존재하지 않는 userId입니다.: " + userId));
+        User user = userRepository.findById(loginUser.id()).orElseThrow(() ->
+                new DataNotFoundException("존재하지 않는 userId입니다.: " + loginUser.id()));
+
+        if (user.isDeleted()) {
+            throw new DataNotFoundException("해당 유저를 찾을 수 없습니다.:" + loginUser.id());
+        }
 
         return UserProfileDto.from(user);
     }
@@ -51,6 +58,11 @@ public class UserService {
         String nickname = request.nickname();
         User user = userRepository.findById(loginUser.id())
                 .orElseThrow(() -> new BadRequestException("존재하지 않는 userId입니다."));
+
+        if (user.isDeleted()) {
+            throw new DataNotFoundException("해당 유저를 찾을 수 없습니다.:" + loginUser.id());
+        }
+
         if (user.getNickname().equals(nickname) || userRepository.existsByNickname(nickname)) {
             throw new BadRequestException("중복된 닉네임이 존재합니다.");
         }
@@ -68,8 +80,15 @@ public class UserService {
         User user = userRepository.findById(loginUser.id())
                 .orElseThrow(() -> new EntityNotFoundException("올바른 ID가 존재하지 않습니다."));
 
+        if (user.isDeleted()) {
+            throw new DataNotFoundException("해당 유저를 찾을 수 없습니다.:" + loginUser.id());
+        }
+
         oauthMemberClientComposite.deleteAccount(user.getOauthId().oauthServer(), user.getRefreshToken());
 
-        userRepository.delete(user);
+        user.deleteUser(CommonUtil.generateRandomDeleteNickname(), CommonUtil.generateRandomDeleteEmail(),
+                new OauthId(CommonUtil.generateRandomDeleteOauthId(), user.getOauthId().oauthServer()));
+
+        formService.deleteFormByUser(user);
     }
 }
